% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_vertical.R
\name{flip_vertical}
\alias{flip_vertical}
\title{A Vertical Flipper}
\usage{
flip_vertical(nodes, y_nudge)
}
\description{
Flip the tree vertically, to reverse the vertical order of nodes
}
\examples{
flip_vertical(nodes, .025)
}
\keyword{flip,}
\keyword{vertical}
\keyword{y,}
