% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midpoint.R
\name{midpoint}
\alias{midpoint}
\title{A Midpoint Function}
\usage{
midpoint(x1, x2)
}
\description{
Compute the midpoints between the corresponding elements of x1 and x2
}
\examples{
midpoint(c(1, 2), c(3, 4))
}
\keyword{midpoint}
