% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_y.R
\name{place_descendents}
\alias{place_descendents}
\title{A Descendent y-coordinate Computer}
\usage{
place_descendents(nodes, parent_id, top, bottom)
}
\description{
Compute the y-coordinates for all descendents of the present node
}
\examples{
place_descendents(nodes, 3, .5, 0)
}
\keyword{coordinate,}
\keyword{descendent}
\keyword{y,}
