% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_nodes.R
\name{place_nodes}
\alias{place_nodes}
\title{A Label Placer}
\usage{
place_nodes(nodes)
}
\description{
Compute the x- and y-coordinate for each label
}
\examples{
place_nodes(nodes)
}
\keyword{coordinate}
\keyword{x,}
\keyword{y,}
