% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_segment_labels.R
\name{place_segment_labels}
\alias{place_segment_labels}
\title{A Segment-Label Placer}
\usage{
place_segment_labels(nodes, y_nudge, aspect_ratio)
}
\description{
Compute the coordinates and angle each line segment's label/probability
}
\examples{
place_segment_labels(nodes, .05, 1.4)
}
\keyword{angle,}
\keyword{line}
\keyword{segment}
\keyword{segment,}
\keyword{x,}
\keyword{y,}
