% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_segments.R
\name{place_segments}
\alias{place_segments}
\title{A Segment Placer}
\usage{
place_segments(nodes)
}
\description{
Compute the x, y, xend, and yend for each line segment connecting labels in the tree
}
\examples{
place_segments(nodes)
}
\keyword{coordinate,}
\keyword{line}
\keyword{segment}
\keyword{segment,}
\keyword{x,}
\keyword{xend,}
\keyword{y,}
\keyword{yend,}
