% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathS3.R
\name{plot.path}
\alias{plot.path}
\title{S3 plot function for a given path object}
\usage{
\method{plot}{path}(path, ...)
}
\arguments{
\item{path}{Path object to be plotted}
}
\description{
S3 plot function for a given path object
}
