\name{PTree-package}
\alias{PTree-package}
\alias{PTree}
\docType{package}
\title{
\packageTitle{PTree}
}
\description{
\packageDescription{PTree}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PTree}
\packageIndices{PTree}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
For the paper "Growing the efficient frontier on panel trees"
}
\author{
\packageAuthor{PTree}
Jingyu He, Xin He.
email: xin.he@ustc.edu.cn
Maintainer: \packageMaintainer{PTree}
}
\references{
    @article{cong2024growing,
    title={Growing the efficient frontier on panel trees},
    author={Cong, Lin William and Feng, Guanhao and He, Jingyu and He, Xin},
    journal={Journal of Financial Economics, forthcoming},
    year={2024}
    }
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
See ./replication-JFE
}
