% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tree.R
\name{compute_tree}
\alias{compute_tree}
\title{A Tree Computer}
\usage{
compute_tree(nodes, round = 2, ynudge = 0.025, aspect_ratio = 1.2)
}
\description{
Compute all information (coordinates, angles) needed to draw a tree in a plotting package, e.g. ggplot2
}
\examples{
compute_tree(nodes, 2, .05, 1.4)
}
\keyword{angle,}
\keyword{line}
\keyword{segment}
\keyword{segment,}
\keyword{x,}
\keyword{y,}
