% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_tree.R
\name{draw_tree}
\alias{draw_tree}
\title{A Probability Tree Plotter}
\usage{
draw_tree(
  nodes,
  padding = text_padding_default(),
  leaf_p_fontface = "bold",
  show_leaf_probs = TRUE
)
}
\description{
Draw a probability tree using ggplot2
}
\examples{
draw_tree(nodes)
}
\keyword{angle,}
\keyword{line}
\keyword{segment}
\keyword{segment,}
\keyword{x,}
\keyword{y,}
