% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_tree.R
\name{p_tree}
\alias{p_tree}
\title{A Tree Planner & Plotter}
\usage{
p_tree(nodes)
}
\description{
Lay out a probability tree, then draw it using ggplot2
}
\examples{
p_tree(nodes)
}
\keyword{compute,}
\keyword{draw,}
\keyword{tree}
