% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_labels.R
\name{place_labels}
\alias{place_labels}
\title{A Label Placer}
\usage{
place_labels(nodes)
}
\description{
Compute the x- and y-coordinate for each label
}
\examples{
place_labels(nodes)
}
\keyword{coordinate}
\keyword{x,}
\keyword{y,}
